refactor: list component [#124]
